requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
aiosqlite==0.19.0
annotated-types==0.5.0
async-timeout==4.0.2
attrs==23.1.0
charset-normalizer==3.1.0
confz==2.0.1
discord.py==2.3.1
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
pydantic==2.3.0
pydantic_core==2.6.3
python-dotenv==1.0.0
PyYAML==6.0.1
toml==0.10.2
typing_extensions==4.7.1
yarl==1.9.2
